pyproject.toml
[project]
name = "gymbudsapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.10"
dependencies = [
    "poetry (>=2.1.1,<3.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "ngrok (>=1.4.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
